requirements.txt
mss
pillow
pandas
keyboard
requests
opencv-python
pygame
torch
torchvision
scikit-learn
matplotlib
pydirectinput
vgamepad